feat: :sparkles: feat and fixes
fix modules imports, package json for migrations
