+++ package.json
@@ -2,15 +2,14 @@
   "name": "backend-challenge",
   "version": "1.0.0",
   "description": "## Descripción",
-  "main": "index.js",
-  "type": "module",
+  "main": "main.js",
   "scripts": {
     "build": "tsc",
-    "dev": "npm run build && tsx watch src/index.ts",
-    "start": "node dist/index.js",
+    "dev": "npm run build && tsx watch src/main.ts",
+    "start": "node dist/src/main.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "eslint",
-    "typeorm": "typeorm-ts-node-esm -d ./src/db.ts",
+    "typeorm": "tsx ./node_modules/typeorm/cli.js migration:generate -d src/db.ts",
     "m:gen": "npm run typeorm migration:generate",
     "m:run": "npm run typeorm migration:run"
   },
