+++ src/components/property/property.entity.ts
@@ -11,10 +11,10 @@ import {
   Relation,
   Index,
 } from 'typeorm';
-import { User } from '../user/user.entity.js';
-import { Advertisement } from '../advertisement/advertisement.entity.js';
-import { IGeoLocation, IProperty, PropertySectorEnum, TPropertySectorEnum } from './property.dto.js';
-import { Transaction } from '../transaction/transaction.entity.js';
+import { User } from '../user/user.entity';
+import { Advertisement } from '../advertisement/advertisement.entity';
+import { IGeoLocation, IProperty, PropertySectorEnum, TPropertySectorEnum } from './property.dto';
+import { Transaction } from '../transaction/transaction.entity';
 
 @Entity()
 export class Property extends BaseEntity implements IProperty {
