+++ src/components/transaction/transaction.dto.ts
@@ -1,7 +1,7 @@
-import { Property } from "../property/property.entity.js";
-import { Advertisement } from "../advertisement/advertisement.entity.js";
-import { IBaseEntityType } from "../user/user.dto.js";
-import { User } from "../user/user.entity.js";
+import { Property } from "../property/property.entity";
+import { Advertisement } from "../advertisement/advertisement.entity";
+import { IBaseEntityType } from "../user/user.dto";
+import { User } from "../user/user.entity";
 
 export enum TransactionTypeEnum {
   SALE_PURCHASE = 'sale_purchase',
