+++ vercel.json
@@ -1,5 +1,6 @@
 {
-  "rewrites": [{ "source": "/(.*)", "destination": "/" }],
+  "version": 2,
+  "rewrites": [{ "source": "/(.*)", "destination": "/api" }],
   "installCommand": "npm install",
   "buildCommand": "npm run build"
 }
